+++ spec/models/categ.rb
@@ -5,4 +5,9 @@ class Categ < ActiveRecord::Base
   self.primary_key = :cat_id
 
   has_many :subcategs, :foreign_key => :fk_subcat_id
+
+  def cat_id
+    # required for Rails 3.2 compatibility
+    read_attribute(:cat_id)
+  end
 end