+++ spec/models/company.rb
@@ -6,8 +6,6 @@ class Company < ActiveRecord::Base
   belongs_to :parent, :class_name => 'Company', :foreign_key => 'parent_id'
   has_many :children, :class_name => 'Company', :foreign_key => 'parent_id'
 
-  has_many :company_access_levels
-  has_many :recruiters, through: :company_access_levels
-
   counter_culture :parent, :column_name => :children_count
+
 end